requirements.txt
fastapi
uvicorn
textblob
streamlit
requests
python-dotenv